requirements_test.txt
# Testing dependencies
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
coverage>=7.2.0

# Existing dependencies
PySide6>=6.6.0
reportlab>=4.0.0
pyzbar>=0.1.9
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
numpy>=1.24.0
